+++ carp_mobile_sensing/lib/sampling_packages/sensors/sensor_probes.dart
@@ -21,6 +21,72 @@ class UserAccelerometerProbe extends StreamProbe {
       Measurement.fromData(Acceleration(x: event.x, y: event.y, z: event.z)));
 }
 
+/// A probe collecting average accelerometer data collected over a sampling
+/// period. Configured with a [PeriodicSamplingConfiguration] configuration.
+class AverageUserAccelerometerProbe extends BufferingPeriodicStreamProbe {
+  List<UserAccelerometerEvent> userAccelerometerEventList = [];
+  int sensorStartTime = 0;
+  int? sensorEndTime;
+
+  @override
+  Stream<dynamic> get bufferingStream => userAccelerometerEvents;
+
+  @override
+  Future<Measurement?> getMeasurement() async {
+    if (userAccelerometerEventList.isEmpty) return null;
+
+    // compute averages of accelerometer
+    // xm, ym, zm: normal averages (or means)
+    // xms, yms, zms: averages of squared values
+    // n: number of values included
+    int n = userAccelerometerEventList.length;
+    double xms = 0, yms = 0, zms = 0;
+    double xm = 0, ym = 0, zm = 0;
+
+    for (UserAccelerometerEvent event in (userAccelerometerEventList)) {
+      xms += (event.x) * (event.x);
+      xm += event.x;
+      yms += (event.y) * (event.y);
+      ym += event.y;
+      zms += (event.z) * (event.z);
+      zm += event.z;
+    }
+
+    xms = xms / n;
+    xm = xm / n;
+    yms = yms / n;
+    ym = ym / n;
+    zms = zms / n;
+    zm = zm / n;
+
+    var data = AverageAccelerometer(
+        xm: xm, ym: ym, zm: zm, xms: xms, yms: yms, zms: zms, n: n);
+
+    return Measurement(
+        sensorStartTime: sensorStartTime,
+        sensorEndTime: sensorEndTime,
+        data: data);
+  }
+
+  @override
+  void onSamplingStart() {
+    sensorStartTime = DateTime.now().microsecondsSinceEpoch;
+    userAccelerometerEventList.clear();
+  }
+
+  @override
+  void onSamplingEnd() {
+    sensorEndTime = DateTime.now().microsecondsSinceEpoch;
+  }
+
+  @override
+  void onSamplingData(event) {
+    if (event is UserAccelerometerEvent) {
+      userAccelerometerEventList.add(event);
+    }
+  }
+}
+
 /// A probe collecting raw data from the gyroscope.
 class GyroscopeProbe extends StreamProbe {
   @override